requirements.txt
langchain
langgraph
tavily-python
BeautifulSoup4
selenium
webdriver-manager
openai
flask
json5
Flask-CORS
langchain_openai
python-dotenv
requests